refactor : 폴더 기준 라우팅으로 폴더구조 변경
