+++ pages/folder/index.tsx
