+++ pages/signin/index.tsx
@@ -1,38 +1,35 @@
 import SignHeader from "@/src/components/Sign/SignHeader";
 import SocialSign from "@/src/components/Sign/SocialSign";
-import style from "@/pages/sign.module.css";
+import style from "@/pages/signin/sign.module.css";
 import { checkEmailError, checkPwdError } from "@/src/util/handleSignError";
 import { FormEvent, useRef, useState } from "react";
-import { postSignUp } from "@/src/api/postSignUp";
+import { postSignIn } from "@/src/api/postSignIn";
 import Image from "next/image";
 import EyeOffIcon from "@/src/assets/img/eye-off.svg";
 import EyeOnIcon from "@/src/assets/img/eye-on.svg";
-import { isEmailUnique } from "@/src/api/isEmailUnique";
 import clsx from "clsx";
 
-function SignUpPage() {
+function SignInPage() {
   const [emailErrorMessage, setEmailErrorMessage] = useState("");
   const [pwdErrorMessage, setPwdErrorMessage] = useState("");
-  const [pwdCheckErrorMessage, setPwdCheckErrorMessage] = useState("");
   const [pwdType, setPwdType] = useState("password");
-  const [pwdCheckType, setPwdCheckType] = useState("password");
   const emailInput = useRef<any>("");
   const pwdInput = useRef<any>("");
-  const pwdCheckInput = useRef<any>("");
 
   const handleButtonClick = async (e: FormEvent) => {
     e.preventDefault();
 
-    if (
-      (await handleEmailError()) &&
-      handlePwdError() &&
-      handlePwdCheckError()
-    ) {
+    if (handleEmailError() && handlePwdError()) {
       //이메일, 비밀번호가 유효한 값일 때만 api호출
-      postSignUp({
+      const result = postSignIn({
         email: emailInput.current,
         password: pwdInput.current,
       });
+      const data = await result;
+      if (data) {
+        setEmailErrorMessage("이메일을 확인해주세요");
+        setPwdErrorMessage("비밀번호를 확인해주세요");
+      }
     }
   };
 
@@ -43,48 +40,28 @@ function SignUpPage() {
     inputRef.current = e.target.value;
   };
 
-  const handleEmailError = async () => {
-    let isValid = false;
+  const handleEmailError = () => {
     const errorMessage = checkEmailError(emailInput.current);
     setEmailErrorMessage(errorMessage);
-    if (!errorMessage) {
-      //이메일이 빈값이 아니고 올바른 형태일 때 중복체크, 그 전에는 하지 않음
-      if (await isEmailUnique(emailInput.current)) {
-        setEmailErrorMessage("");
-        isValid = true;
-      } else setEmailErrorMessage("이미 사용 중인 이메일입니다.");
-    }
-
-    return isValid;
+    return !Boolean(errorMessage);
   };
 
   const handlePwdError = () => {
-    handlePwdCheckError();
     const errorMessage = checkPwdError(pwdInput.current);
     setPwdErrorMessage(errorMessage);
     return !Boolean(errorMessage);
   };
-  const handlePwdCheckError = () => {
-    if (pwdCheckInput.current != pwdInput.current) {
-      setPwdCheckErrorMessage("비밀번호가 일치하지 않습니다.");
-      return false;
-    }
-    setPwdCheckErrorMessage("");
-    return true;
-  };
-  const handlePwdEyeClick = (
-    type: string,
-    setType: React.Dispatch<React.SetStateAction<string>>
-  ) => {
-    setType(type === "email" ? "password" : "email");
+
+  const handlePwdEyeClick = () => {
+    setPwdType(pwdType === "email" ? "password" : "email");
   };
   return (
     <div className={style.root}>
       <div className={style.container}>
         <SignHeader
-          message="이미 회원이신가요? "
-          href="/signin"
-          linkMessage="로그인 하기"
+          message="회원이 아니신가요? "
+          href="/signup"
+          linkMessage="회원 가입하기"
         />
         <form className={style.form}>
           <div>
@@ -123,43 +100,19 @@ function SignUpPage() {
                 src={pwdType === "password" ? EyeOffIcon : EyeOnIcon}
                 alt="pwd eye"
                 className={style.pwdEye}
-                onClick={() => handlePwdEyeClick(pwdType, setPwdType)}
+                onClick={handlePwdEyeClick}
               />
             </div>
             <p className={style.errorMessage}>{pwdErrorMessage}</p>
           </div>
-          <div>
-            <label htmlFor="pwd_check_input" className={style.label}>
-              비밀번호 확인
-            </label>
-            <div className={style.inputWrapper}>
-              <input
-                className={clsx(style.input, {
-                  [style.redBox]: pwdCheckErrorMessage,
-                })}
-                id="pwd_check_input"
-                type={pwdCheckType}
-                placeholder="비밀번호를 입력해주세요"
-                onChange={(e) => handleInputChange(e, pwdCheckInput)}
-                onBlur={handlePwdCheckError}
-              />
-              <Image
-                src={pwdCheckType === "password" ? EyeOffIcon : EyeOnIcon}
-                alt="pwd eye"
-                className={style.pwdEye}
-                onClick={() => handlePwdEyeClick(pwdCheckType, setPwdCheckType)}
-              />
-            </div>
-            <p className={style.errorMessage}>{pwdCheckErrorMessage}</p>
-          </div>
           <button className={style.formButton} onClick={handleButtonClick}>
-            회원가입
+            로그인
           </button>
         </form>
-        <SocialSign message="다른 방식으로 가입하기" />
+        <SocialSign message="소셜 로그인" />
       </div>
     </div>
   );
 }
 
-export default SignUpPage;
+export default SignInPage;
